Nested arrays (parking) (#197)
* add experiment project, build hierarchical schema

* build out schema path

* nested non-repeated records are read, but not collected back to dataset

* reading nested structures into rows

* nested structures read in rows

* int fix

* new AddColumnar was only ever adding first row

* always return definition for merging, as null count can differ between pages

* update featues page

* generate test file

* fix test project

* schema is read out properly for nested arrays, read out repetition levels

* small fixes

* slight refactoring - moving out schema building code to a separate class; general comments

* schema & schemaelement improvements; generating test files for all nesting permutations

* pretty looking Schema.ToString()

* repetitions are grouped

* schema parsed to more understandable format, but rep and def levels are ruined i think

* slightly better way of pre-calculating column levels

* parking work on nested structures, making sure code is backward compatible and all tests are passing

* update docs

+++ src/Parquet/ReaderOptions.cs
@@ -19,18 +19,12 @@ namespace Parquet
       /// <summary>
       /// Gets or sets the offset.
       /// </summary>
-      /// <value>
-      /// The offset.
-      /// </value>
       public long Offset { get; set; }
 
       /// <summary>
       /// Gets or sets the count.
       /// </summary>
-      /// <value>
-      /// The count.
-      /// </value>
-      public long Count { get; set; }
+      public int Count { get; set; }
 
       internal void Validate()
       {
